Move Material buttons "can't reserve" & "loading" into generic subfolder
All of a sudden there were too many material buttons in one folder. Now
they are divided in generic, physical and online folders.
