+++ src/components/material/material-buttons/generic/MaterialButtonCantReserve.tsx
@@ -1,11 +1,13 @@
 import * as React from "react";
 import { FC } from "react";
-import { Button } from "../../Buttons/Button";
+import { useText } from "../../../../core/utils/text";
+import { Button } from "../../../Buttons/Button";
 
 const MaterialButtonCantReserve: FC = () => {
+  const t = useText();
   return (
     <Button
-      label="Kan ej reserveres"
+      label={t("cantReserveText")}
       buttonType="none"
       variant="filled"
       disabled
