+++ src/components/material/material-buttons/generic/MaterialButtonLoading.tsx
@@ -1,11 +1,13 @@
 import * as React from "react";
 import { FC } from "react";
-import { Button } from "../../Buttons/Button";
+import { useText } from "../../../../core/utils/text";
+import { Button } from "../../../Buttons/Button";
 
 const MaterialButtonLoading: FC = () => {
+  const t = useText();
   return (
     <Button
-      label="Loading"
+      label={t("loadingText")}
       buttonType="none"
       variant="filled"
       disabled
